Refactor ReplicationDocumentSender out of OutgoingReplicationHandler -> better maintainable code (RavenDB-5585)
